% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controlCountreg}
\alias{controlCountreg}
\title{Controls for zerotrunc class}
\usage{
controlCountreg(
  trcount = 0,
  cores = 1L,
  alpha = 0.05,
  B = 500,
  fittingMethod = c("IRLS", "optim"),
  bootstrapFitcontrol = singleRcapture::controlMethod(epsilon = 0.001, maxiter = 20,
    optimMethod = "Nelder-Mead", silent = TRUE),
  traceBootstrapSize = FALSE,
  bootstrapVisualTrace = FALSE,
  confType = c("percentilic", "normal", "basic"),
  bootType = c("parametric", "semiparametric", "nonparametric"),
  keepbootStat = TRUE
)
}
\arguments{
\item{trcount}{a}

\item{cores}{a}

\item{alpha}{a}

\item{B}{a}

\item{fittingMethod}{a}

\item{bootstrapFitcontrol}{a}

\item{traceBootstrapSize}{a}

\item{bootstrapVisualTrace}{a}

\item{confType}{a}

\item{bootType}{a}

\item{keepbootStat}{a}
}
\value{
a
}
\description{
Controls for zerotrunc class
}
