% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controlEstPopVglm}
\alias{controlEstPopVglm}
\title{Control parameters for \code{vglm} class}
\usage{
controlEstPopVglm(
  alpha = 0.05,
  bootType = c("parametric", "semiparametric", "nonparametric"),
  B = 500,
  confType = c("percentilic", "normal", "basic"),
  keepbootStat = TRUE,
  traceBootstrapSize = FALSE,
  bootstrapVisualTrace = FALSE,
  bootstrapFitcontrol = VGAM::vglm.control(epsilon = 0.001, maxit = 10, noWarning = TRUE),
  sd = c("sqrtVar", "normalMVUE"),
  cores = 1L
)
}
\arguments{
\item{alpha}{Numeric value indicating significance level. By default \code{.05}.}

\item{bootType}{Type of bootstrap by default \code{parametric}. For more detail
see \code{\link[singleRcapture:estimatePopsize]{singleRcapture::estimatePopsize()}}.}

\item{B}{Number of bootstrap samples to be drawn. By default \code{500}.}

\item{confType}{Type of confidence interval to use for bootstrap variance
estimation.}

\item{keepbootStat}{Boolean value indicating whether to keep statistics from
bootstrap samples.}

\item{traceBootstrapSize}{Boolean value indicating whether to print sample
sizes (and estimate population sizes) on each iteration of bootstrap
algorithm. Only matters when \code{cores = 1}.}

\item{bootstrapVisualTrace}{Logical value indicating whether to make an
in real time plot of estimated statistics for \code{cores = 1} or whether
to make progress bar when \code{cores > 1}.}

\item{bootstrapFitcontrol}{A \code{\link[VGAM:vglm.control]{VGAM::vglm.control()}} type object controlling
behaviour of \code{vglm.fit} function used for fitting.}

\item{sd}{Type of standard error estimate.}

\item{cores}{Number of cores to use be default \code{1}.}
}
\value{
A list with desired control specifications.
}
\description{
Control parameters for \code{vglm} class
}
\seealso{
\code{\link[singleRcapture:estimatePopsize]{singleRcapture::estimatePopsize()}} for details on bootstrap
\code{\link[=estimatePopsize.vglm]{estimatePopsize.vglm()}} for details on computation
\code{\link[singleRcapture:controlPopVar]{singleRcapture::controlPopVar()}} for comparison with \code{singleRcapture}
\code{\link[VGAM:vglm.control]{VGAM::vglm.control()}} for fitting controls in bootstrap
\code{\link[VGAM:notdocumentedyet]{VGAM::vglm.fit()}} for information on fitting algorithm
}
