% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R, R/controls.R
\name{controls}
\alias{controls}
\alias{controlEstPopCountreg}
\alias{controlEstPopVglm}
\title{Control parameters for \code{vgam}, \code{vglm}, \code{countreg} class}
\usage{
controlEstPopCountreg(
  trcount = 0,
  cores = 1L,
  alpha = 0.05,
  B = 500,
  fittingMethod = c("IRLS", "optim"),
  bootstrapFitcontrol = singleRcapture::controlMethod(epsilon = 0.001, maxiter = 20,
    optimMethod = "Nelder-Mead", silent = TRUE),
  traceBootstrapSize = FALSE,
  bootstrapVisualTrace = FALSE,
  confType = c("percentilic", "normal", "basic"),
  bootType = c("parametric", "semiparametric", "nonparametric"),
  keepbootStat = TRUE
)

controlEstPopVglm(
  alpha = 0.05,
  bootType = c("parametric", "semiparametric", "nonparametric"),
  B = 500,
  confType = c("percentilic", "normal", "basic"),
  keepbootStat = TRUE,
  traceBootstrapSize = FALSE,
  bootstrapVisualTrace = FALSE,
  bootstrapFitcontrol = VGAM::vgam.control(epsilon = 0.001, maxit = 10, noWarning = TRUE),
  sd = c("sqrtVar", "normalMVUE"),
  cores = 1L,
  data = NULL
)

controlEstPopVglm(
  alpha = 0.05,
  bootType = c("parametric", "semiparametric", "nonparametric"),
  B = 500,
  confType = c("percentilic", "normal", "basic"),
  keepbootStat = TRUE,
  traceBootstrapSize = FALSE,
  bootstrapVisualTrace = FALSE,
  bootstrapFitcontrol = VGAM::vgam.control(epsilon = 0.001, maxit = 10, noWarning = TRUE),
  sd = c("sqrtVar", "normalMVUE"),
  cores = 1L,
  data = NULL
)
}
\arguments{
\item{trcount}{todo}

\item{cores}{Number of cores to use be default \code{1}.}

\item{alpha}{Numeric value indicating significance level. By default \code{.05}.}

\item{B}{Number of bootstrap samples to be drawn. By default \code{500}.}

\item{fittingMethod}{For \code{countreg} class which fitting algorithm from
\code{singleRcapture} to use (bootstrap for \code{countreg} class calls
bootstrap from \code{singleRcapture}).}

\item{bootstrapFitcontrol}{A \code{\link[VGAM:vglm.control]{VGAM::vglm.control()}} type object controlling
behaviour of \code{vglm.fit} function used for fitting.}

\item{traceBootstrapSize}{Boolean value indicating whether to print sample
sizes (and estimate population sizes) on each iteration of bootstrap
algorithm. Only matters when \code{cores = 1}.}

\item{bootstrapVisualTrace}{Logical value indicating whether to make an
in real time plot of estimated statistics for \code{cores = 1} or whether
to make progress bar when \code{cores > 1}.}

\item{confType}{Type of confidence interval to use for bootstrap variance
estimation.}

\item{bootType}{Type of bootstrap by default \code{parametric}. For more detail
see \code{\link[singleRcapture:estimatePopsize]{singleRcapture::estimatePopsize()}}.}

\item{keepbootStat}{Boolean value indicating whether to keep statistics from
bootstrap samples.}

\item{sd}{Type of standard error estimate.}

\item{data}{for \code{vgam} class only, a original \code{data.frame} object
on which model was fitted. Only required for bootstrap.}
}
\value{
A list with desired control specifications.
}
\description{
Control parameters for \code{vgam}, \code{vglm}, \code{countreg} class
}
\seealso{
\code{\link[singleRcapture:estimatePopsize]{singleRcapture::estimatePopsize()}} for details on bootstrap and fitting algorithms
\code{\link[=estimatePopsize.vglm]{estimatePopsize.vglm()}} for details on computation
\code{\link[singleRcapture:controlPopVar]{singleRcapture::controlPopVar()}} for comparison with \code{singleRcapture}
\code{\link[VGAM:vglm.control]{VGAM::vglm.control()}} for fitting controls in bootstrap
\code{\link[VGAM:notdocumentedyet]{VGAM::vglm.fit()}} for information on fitting algorithm
}
