% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R, R/dfpopsize.R
\name{dfpop}
\alias{dfpop}
\alias{dfpopsize.singleRadditive}
\alias{dfpopsize.singleRforeign}
\title{Leave one out diagnostic on population size}
\usage{
\method{dfpopsize}{singleRadditive}(model, data = NULL, cores = 1L, trace = FALSE, ...)

\method{dfpopsize}{singleRforeign}(model, dfbeta = NULL, ...)
}
\arguments{
\item{model}{object of \code{singleRforeign} class.}

\item{data}{A \code{data.frame} provided at call to the fitting function. This is
usually not needed since the function first tries to find this \code{data.frame}
manually. This argument only needs to be provided when the original object was
removed or renamed in the global environment.}

\item{cores}{Number of processor cores to be used for computing the population
size estimates for additive models. Unfortunately for \code{VGAM} class
we need to fit a new model ater removing each unit in the data.}

\item{trace}{logical value specifying whether to track the results. Only
applicable if a single core is used.}

\item{...}{b}

\item{dfbeta}{if \code{dfbeta} was already obtained it is possible to pass
them into function so that they need not be computed for the second time.}
}
\value{
A numeric vector with the differences between the estimated population
size and the estimated population size after removing a unit.
}
\description{
Leave one out diagnostic on population size
}
