% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dfpopsize.singleRadditive}
\alias{dfpopsize.singleRadditive}
\title{AA}
\usage{
\method{dfpopsize}{singleRadditive}(model, cores = 1L, ...)
}
\arguments{
\item{model}{a}

\item{cores}{a}

\item{...}{b}
}
\value{
a
}
\description{
AA
}
