% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dfpopsize.singleRforeign}
\alias{dfpopsize.singleRforeign}
\title{AAA}
\usage{
\method{dfpopsize}{singleRforeign}(model, dfbeta = NULL, ...)
}
\arguments{
\item{model}{a}

\item{dfbeta}{a}

\item{...}{a}
}
\value{
a
}
\description{
AAA
}
