% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{marginalFreqVglm}
\alias{marginalFreqVglm}
\title{AA}
\usage{
marginalFreqVglm(object, includezeros = TRUE, range, ...)
}
\arguments{
\item{object}{A}

\item{includezeros}{A}

\item{range}{A}

\item{...}{A}
}
\value{
A
}
\description{
AA
}
