% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R, R/marginal.R
\name{marginals}
\alias{marginals}
\alias{marginalFreqVglm}
\alias{marginalFreqCountreg}
\title{Observed and fitted marginal Frequencies}
\usage{
marginalFreqVglm(object, includezeros = TRUE, range, ...)

marginalFreqCountreg(object, includezeros = TRUE, range, ...)
}
\arguments{
\item{object}{object of \code{singleR} class.}

\item{includezeros}{logical value indicating whether to include one counts in the zero-one truncated models.}

\item{range}{optional argument specifying range of selected Y values.}

\item{...}{currently does nothing.}
}
\value{
A list with observed name of the fitted model family degrees of freedom and observed and fitted marginal frequencies.
}
\description{
Observed and fitted marginal Frequencies
}
\seealso{
\code{\link[singleRcapture:marginalFreq]{singleRcapture::marginalFreq()}} -- for original \code{singleRcapture} implementation. \code{\link[=estimatePopsize.vgam]{estimatePopsize.vgam()}}
}
\author{
Piotr Chlebicki
}
