% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.singleRforeign}
\alias{plot.singleRforeign}
\title{Title}
\usage{
\method{plot}{singleRforeign}(
  x,
  plotType = c("bootHist", "marginal", "rootogram", "dfpopContr", "dfpopBox", "strata"),
  confIntStrata = c("normal", "logNormal"),
  histKernels = TRUE,
  dfpop,
  ...
)
}
\arguments{
\item{x}{A}

\item{plotType}{A}

\item{confIntStrata}{A}

\item{histKernels}{A}

\item{dfpop}{A}

\item{...}{A}
}
\value{
A
}
\description{
Title
}
