% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyPopsize.R
\name{stratifyPopsize.singleRforeign}
\alias{stratifyPopsize.singleRforeign}
\title{AAA}
\usage{
\method{stratifyPopsize}{singleRforeign}(object, stratas, alpha, cov = NULL, ...)
}
\arguments{
\item{object}{a}

\item{stratas}{a}

\item{alpha}{description}

\item{cov}{a}

\item{...}{a}
}
\value{
A
}
\description{
AA
}
\details{
aa
}
\seealso{
\code{\link[singleRcapture:stratifyPopsize]{singleRcapture::stratifyPopsize()}} \code{\link[=estimatePopsize]{estimatePopsize()}}
}
