% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.singleRforeign}
\alias{summary.singleRforeign}
\title{AAA}
\usage{
\method{summary}{singleRforeign}(object, popSizeEst, ...)
}
\arguments{
\item{object}{a}

\item{popSizeEst}{a}

\item{...}{a}
}
\description{
AA
}
\details{
aa
}
