% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controlEstPopVglm}
\alias{controlEstPopVglm}
\title{Title}
\usage{
controlEstPopVglm(
  alpha = 0.05,
  bootType = c("parametric", "semiparametric", "nonparametric"),
  B = 500,
  confType = c("percentilic", "normal", "basic"),
  keepbootStat = TRUE,
  traceBootstrapSize = FALSE,
  bootstrapVisualTrace = FALSE,
  bootstrapFitcontrol = VGAM::vglm.control(epsilon = 0.001, maxit = 10, noWarning = TRUE),
  sd = c("sqrtVar", "normalMVUE"),
  cores = 1L
)
}
\arguments{
\item{alpha}{a}

\item{bootType}{a}

\item{B}{a}

\item{confType}{a}

\item{keepbootStat}{a}

\item{traceBootstrapSize}{a}

\item{bootstrapVisualTrace}{a}

\item{bootstrapFitcontrol}{a}

\item{sd}{a}

\item{cores}{a}
}
\value{
a
}
\description{
Title
}
