% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R, R/dfpopsize.R
\name{dfpop}
\alias{dfpop}
\alias{dfpopsize.singleRadditive}
\alias{dfpopsize.singleRforeign}
\title{AA}
\usage{
\method{dfpopsize}{singleRadditive}(model, data, cores = 1L, trace = FALSE, ...)

\method{dfpopsize}{singleRforeign}(model, dfbeta = NULL, ...)
}
\arguments{
\item{model}{a}

\item{cores}{a}

\item{trace}{asd}

\item{...}{b}

\item{dfbeta}{a}
}
\value{
a
}
\description{
AA
}
