% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R, R/estimatePopsizevglm.R
\name{foreignMethods}
\alias{foreignMethods}
\alias{estimatePopsize.vglm}
\alias{estimatePopsize.vgam}
\title{TODO}
\usage{
\method{estimatePopsize}{vglm}(
  formula,
  subset = NULL,
  naAction = NULL,
  popVar = c("analytic", "bootstrap"),
  controlForeign = NULL,
  derivFunc = NULL,
  ...
)

\method{estimatePopsize}{vgam}(
  formula,
  subset = NULL,
  naAction = NULL,
  popVar = c("analytic", "bootstrap"),
  controlForeign = NULL,
  derivFunc = NULL,
  ...
)
}
\arguments{
\item{formula}{A}

\item{subset}{A}

\item{naAction}{A}

\item{popVar}{A}

\item{controlForeign}{A}

\item{derivFunc}{A}

\item{...}{A}
}
\description{
TODO
}
\details{
\loadmathjax

TODO
}
