% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.singleRforeign}
\alias{plot.singleRforeign}
\title{Diagnostic plots for population size estimation.}
\usage{
\method{plot}{singleRforeign}(
  x,
  plotType = c("bootHist", "marginal", "rootogram", "dfpopContr", "dfpopBox", "strata"),
  confIntStrata = c("normal", "logNormal"),
  histKernels = TRUE,
  dfpop,
  ...
)
}
\arguments{
\item{x}{object of \code{singleRforeign} class.}

\item{plotType}{character parameter specifying type of plot to be made,
possible options are
\code{bootHist, marginal, rootogram, dfpopContr, dfpopBox, strata}
and their full description may be found in
\code{\link[singleRcapture:plot.singleRStaticCountData]{singleRcapture::plot.singleRStaticCountData()}}.}

\item{confIntStrata}{confidence interval type to use for strata plot.
Currently supported values are \code{normal} and \code{logNormal}.}

\item{histKernels}{logical value indicating whether to add density lines to histogram.}

\item{dfpop}{if \code{\link[=dfpopsize.singleRforeign]{dfpopsize.singleRforeign()}} was already computed it may be
supplied as a argument to this function in order to reduce the computational
time.}

\item{...}{additional arguments passed to plot functions, full description
of which arguments work is present in \code{\link[singleRcapture:plot.singleRStaticCountData]{singleRcapture::plot.singleRStaticCountData()}}.}
}
\value{
No return value only the plot being made.
}
\description{
Diagnostic plots for population size estimation.
}
