% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.singleRforeign}
\alias{summary.singleRforeign}
\title{AAA}
\usage{
\method{summary}{singleRforeign}(object, popSizeEst, ...)
}
\arguments{
\item{object}{object of \code{singleRforeign} class}

\item{popSizeEst}{object of \code{popSizeEstResults} class such as the ones
created by \code{\link[singleRcapture:redoPopEstimation]{singleRcapture::redoPopEstimation()}}. If not specified
population size estimation results will be drawn from the \code{object}.
results}

\item{...}{passed to \code{summary} method on foreign object provided
on call to \code{estimatePopsize}.}
}
\description{
A summary method for \code{singleRforeign} class
}
\details{
Computes summary statistics for population size estimation and calls
summary on a foreign object. The results can be printed either as (default)
\code{print(summary(object))} where both the summary of population size
estimation and of the foreign object will be printed or alternatively the
later statistics can be omitted in printing with the call
\code{print(summary(object), summaryForeign = FALSE)}.
}
